+++ frontend/src/hooks/tanstack-query/station-details/useStationCongestionStatistics.ts
@@ -5,7 +5,8 @@ import { useExternalValue } from '@utils/external-state';
 import { selectedStationIdStore } from '@stores/selectedStationStore';
 import { serverStore } from '@stores/serverStore';
 
-import { ERROR_MESSAGES, SERVERS } from '@constants';
+import { SERVERS } from '@constants';
+import { ERROR_MESSAGES } from '@constants/errorMessages';
 
 import type { CongestionStatistics } from '../../../types';
