+++ frontend/src/mocks/handlers.ts
@@ -2,7 +2,8 @@ import { rest } from 'msw';
 
 import { getSessionStorage, setSessionStorage } from '@utils/storage';
 
-import { ERROR_MESSAGES, SERVERS, SESSION_KEY_REPORTED_STATIONS } from '@constants';
+import { SERVERS, SESSION_KEY_REPORTED_STATIONS } from '@constants';
+import { ERROR_MESSAGES } from '@constants/errorMessages';
 
 import { getCongestionStatistics, getSearchedStations, stations } from './data';
